Remove useless casts

+++ MapEver/src/de/hu_berlin/informatik/spws2014/mapever/largeimageview/LargeImageView.java
@@ -423,8 +423,8 @@ public class LargeImageView extends AppCompatImageView {
         if (point == null) {
             panCenterX = panCenterY = 0;
         } else {
-            panCenterX = (float) point.x;
-            panCenterY = (float) point.y;
+            panCenterX = point.x;
+            panCenterY = point.y;
         }
         update();
     }
